catalog-backend: split out ProcessorCacheManager and add tests
